refactor seperation of concers
