+++ go.mod
@@ -3,3 +3,5 @@ module github.com/jordanOBL/sleepBubble
 go 1.23.0
 
 require github.com/oliveroneill/exponent-server-sdk-golang v0.0.0-20210823140141-d050598be512 // indirect
+
+require github.com/julienschmidt/httprouter v1.3.0
