--- internal/server/middleware.go
@@ -1,1 +1,0 @@
-package server
